requirements.txt
matplotlib==3.10.5
openpyxl==3.1.5
pandas==2.3.1
pip-chill==1.0.3
pipreqs==0.5.0
pyserial==3.5
pyside6==6.9.1
tinycss2==1.4.0
